+++ src/components/portal/__spec__.js
@@ -1,5 +1,6 @@
 import React from 'react';
-import { mount, shallow } from 'enzyme';
+import { mount } from 'enzyme';
+import ReactDOM from 'react-dom';
 import Portal from './portal';
 import Icon from './../icon';
 import Browser from '../../utils/helpers/browser';
@@ -30,7 +31,7 @@ describe('Portal', () => {
 
     it('can be able to access Icon', () => {
       expect(wrapper.find(Icon).length).toBe(1);
-    });    
+    });
 
     it('will mount second portal', () => {
       const wrapper2 = mount(
@@ -43,7 +44,7 @@ describe('Portal', () => {
           />
         </Portal>
       );
-      
+
       expect(document.body.getElementsByClassName('carbon-portal').length).toEqual(2);
       wrapper2.unmount();
     });
@@ -67,11 +68,110 @@ describe('Portal', () => {
     it('to match snapshot ', () => {
       expect(wrapper).toMatchSnapshot();
     });
+  });
 
+  describe('will manager listeners', () => {
+    describe('when NOT given  reposition prop', () => {
+      let parentDiv;
+      beforeEach(() => {
+        spyOn(Browser.getWindow(), 'addEventListener');
+        spyOn(Browser.getWindow(), 'removeEventListener');
+        spyOn(ReactDOM, 'findDOMNode').and.returnValue(parentDiv);
+
+        parentDiv = Browser.getDocument().createElement('div');
+        spyOn(parentDiv, 'addEventListener');
+        spyOn(parentDiv, 'removeEventListener');
+
+        wrapper = mount(
+          <Portal>
+            <Icon
+              tooltipMessage='Test'
+              tooltipAlign='left'
+              tooltipPosition='top'
+              type='tick'
+            />
+          </Portal>
+        );
+      });
+
+      afterEach(() => {
+        wrapper.unmount();
+      });
+
+      it('will NOT add window "resize" listener ', () => {
+        expect(Browser.getWindow().addEventListener).not.toHaveBeenCalledWith('resize');
+      });
+
+      it('will NOT remove window "resize" listener on unnmount', () => {
+        wrapper.unmount();
+        expect(Browser.getWindow().removeEventListener).not.toHaveBeenCalledWith('resize');
+      });
+
+      it('will NOT window "scroll" listener ', () => {
+        expect(parentDiv.addEventListener).not.toHaveBeenCalled();
+      });
+
+      it('will NOT remove "scroll" listener on unnmount', () => {
+        wrapper.unmount();
+        expect(parentDiv.removeEventListener).not.toHaveBeenCalled();
+      });
+    });
+
+    describe('when given reposition prop', () => {
+      let repositionCb;
+      let parentDiv;
+      beforeEach(() => {
+        repositionCb = jasmine.createSpy('onReposition');
+        parentDiv = Browser.getDocument().createElement('div');
+        parentDiv.style.overflow = 'auto';
+
+        spyOn(Browser.getWindow(), 'addEventListener');
+        spyOn(Browser.getWindow(), 'removeEventListener');
+        spyOn(ReactDOM, 'findDOMNode').and.returnValue(parentDiv);
+        spyOn(parentDiv, 'addEventListener');
+        spyOn(parentDiv, 'removeEventListener');
+        wrapper = mount(
+          <Portal onReposition={ repositionCb }>
+            <Icon
+              tooltipMessage='Test'
+              tooltipAlign='left'
+              tooltipPosition='top'
+              type='tick'
+            />
+          </Portal>
+        );
+      });
+
+      afterEach(() => {
+        wrapper.unmount();
+      });
+
+      it('will add window "resize" listener ', () => {
+        expect(Browser.getWindow().addEventListener).toHaveBeenCalledWith('resize', repositionCb);
+      });
+
+      it('will remove "resize" listener on unnmount', () => {
+        wrapper.unmount();
+        expect(Browser.getWindow().removeEventListener).toHaveBeenCalledWith('resize', repositionCb);
+      });
+
+      it('will call window "reposition" callback ', () => {
+        expect(repositionCb).toHaveBeenCalled();
+      });
+
+      it('will add window "scroll" listener ', () => {
+        expect(parentDiv.addEventListener).toHaveBeenCalledWith('scroll', repositionCb);
+      });
+
+      it('will remove "scroll" listener on unnmount', () => {
+        wrapper.unmount();
+        expect(parentDiv.removeEventListener).toHaveBeenCalledWith('scroll', repositionCb);
+      });
+    });
   });
 
   it('mount a <p/> tag as child', () => {
-    const wrapper = mount(
+    mount(
       <Portal>
         <p>john</p>
       </Portal>
@@ -82,7 +182,7 @@ describe('Portal', () => {
 
   it('will NOT mount with no DOM', () => {
     spyOn(Browser, 'getWindow').and.returnValue(undefined);
-    const wrapper = mount(
+    const noDOMWrapper = mount(
       <Portal>
         <Icon
           tooltipMessage='Test'
@@ -92,6 +192,6 @@ describe('Portal', () => {
         />
       </Portal>
       );
-    expect(wrapper.html()).toBe(null);
+    expect(noDOMWrapper.html()).toBe(null);
   });
 });